Create Module 2 Assignment
We were provided with the first two lines of code and told to evaluate whether this function works. I went ahead and included the errors in the code I recieved and then step-by-step fixed each problem. The biggest thing was that whichever variable you assign to be taken into the function is also the local variable or as least having commands that assign aspects of the global variable to their local variable values.